+++ include/thezhe/avec.hpp
@@ -1,3 +1,0 @@
-#pragma once
-#include "avec/Avec.hpp" // IWYU pragma: keep
-//TODO new name