+++ include/thezhe/d2/D2.hpp
@@ -2,79 +2,77 @@
 #pragma once
 #include <array>
 #include <cmath>
-#define AVEC_STD_UNARY_OP(__op)                                                \
-    [[nodiscard]] constexpr Avec __op()                                        \
-        const noexcept(noexcept(std::__op(T())))                               \
+#define D2_STD_UNARY(__op)                                                     \
+    [[nodiscard]] constexpr D2 __op()                                          \
+        const noexcept(noexcept(std::__op(value_t())))                         \
     {                                                                          \
-        Avec retval{};                                                         \
-        for (std::size_t i = 0; i < n; ++i)                                    \
+        D2 retval{};                                                           \
+        for (std::size_t i = 0; i < size_v; ++i)                               \
         {                                                                      \
             retval.data[i] = std::__op(data[i]);                               \
         }                                                                      \
         return retval;                                                         \
     }
-#define AVEC_STD_BINARY_OP(__op)                                               \
-    [[nodiscard]] constexpr Avec __op(Avec b)                                  \
-        const noexcept(noexcept(std::__op(T(), T())))                          \
+#define D2_STD_BINARY(__op)                                                    \
+    [[nodiscard]] constexpr D2 __op(D2 b)                                      \
+        const noexcept(noexcept(std::__op(value_t(), value_t())))              \
     {                                                                          \
-        Avec retval{};                                                         \
-        for (std::size_t i = 0; i < n; ++i)                                    \
+        D2 retval{};                                                           \
+        for (std::size_t i = 0; i < size_v; ++i)                               \
         {                                                                      \
             retval.data[i] = std::__op(data[i], b.data[i]);                    \
         }                                                                      \
         return retval;                                                         \
     }
-#define AVEC_STD_TERNARY_OP(__op)                                              \
-    [[nodiscard]] constexpr Avec __op(Avec b, Avec c)                          \
-        const noexcept(noexcept(std::__op(T(), T(), T())))                     \
+#define D2_STD_TERNARY(__op)                                                   \
+    [[nodiscard]] constexpr D2 __op(D2 b, D2 c)                                \
+        const noexcept(noexcept(std::__op(value_t(), value_t(), value_t())))   \
     {                                                                          \
-        Avec retval{};                                                         \
-        for (std::size_t i = 0; i < n; ++i)                                    \
+        D2 retval{};                                                           \
+        for (std::size_t i = 0; i < size_v; ++i)                               \
         {                                                                      \
             retval.data[i] = std::__op(data[i], b.data[i], c.data[i]);         \
         }                                                                      \
         return retval;                                                         \
     }
-#define AVEC_ARITHMETIC_OP(__op)                                               \
-    constexpr Avec operator __op##=(Avec other) noexcept                       \
+#define D2_ARITHMETIC(__op)                                                    \
+    constexpr D2 operator __op##=(D2 other) noexcept                           \
     {                                                                          \
-        *this = transform(other, [](T a, T b) noexcept { return a __op b; });  \
+        *this = transform(other,                                               \
+                          [](value_t a, value_t b) noexcept                    \
+                          { return a __op b; });                               \
         return *this;                                                          \
     }                                                                          \
-    [[nodiscard]] friend constexpr Avec operator __op(Avec a, Avec b) noexcept \
+    [[nodiscard]] friend constexpr D2 operator __op(D2 a, D2 b) noexcept       \
     {                                                                          \
         a __op## = b;                                                          \
         return a;                                                              \
     }
-namespace thezhe::avec
+namespace thezhe::d2
 {
 /*!
- * @brief Auto-vectorizable `std::array` facade
- * @tparam T Element type. Must be floating-point.
- * @tparam n Number of elements
- * @pre See `thezhe::avec::preconditions`
- * @note Use like a floating-point type
- * @details In other words: pass by value and avoid mixing specializations for
- * best performance
+ * @brief Auto-vectorization-friendly `std::array` facade
+ * @pre See `thezhe::d2::preconditions`
+ * @attention Pass by value
  */
-template<typename T = double, std::size_t n = 2>
-    requires std::is_floating_point_v<T>
-class Avec final
+class D2 final
 {
 public:
-    constexpr Avec(T t) noexcept // NOLINT
+    using value_t = double;
+    static constexpr std::size_t size_v = 2;
+    constexpr D2(value_t t) noexcept // NOLINT
     {
         *this = transform([t]() noexcept { return t; });
     }
     template<typename... Ts>
-    constexpr explicit Avec(Ts... ts) noexcept : data({ ts... })
+    constexpr explicit D2(Ts... ts) noexcept : data({ ts... })
     {
     }
-    AVEC_ARITHMETIC_OP(+)
-    AVEC_ARITHMETIC_OP(-)
-    AVEC_ARITHMETIC_OP(*)
-    AVEC_ARITHMETIC_OP(/)
-    friend constexpr Avec operator-(Avec a) noexcept
+    D2_ARITHMETIC(+)
+    D2_ARITHMETIC(-)
+    D2_ARITHMETIC(*)
+    D2_ARITHMETIC(/)
+    friend constexpr D2 operator-(D2 a) noexcept
     {
         a *= -1;
         return a;
@@ -83,7 +81,7 @@ public:
      * @brief Default comparison operator. Does NOT account for floating-point
      * error.
      */
-    friend constexpr bool operator==(Avec a, Avec b) noexcept = default;
+    friend constexpr bool operator==(D2 a, D2 b) noexcept = default;
     /*!
      * @brief TODO (vec versions)
      * tanh
@@ -95,35 +93,35 @@ public:
      * lerp
      * fmod/ceil/floor/remainder
      */
-    AVEC_STD_UNARY_OP(sqrt)
-    AVEC_STD_UNARY_OP(abs)
-    AVEC_STD_BINARY_OP(min)
-    AVEC_STD_BINARY_OP(max)
-    AVEC_STD_BINARY_OP(tanh)
-    AVEC_STD_BINARY_OP(pow)
-    AVEC_STD_BINARY_OP(tan)
+    D2_STD_UNARY(sqrt)
+    D2_STD_UNARY(abs)
+    D2_STD_BINARY(min)
+    D2_STD_BINARY(max)
+    D2_STD_UNARY(tanh)
+    D2_STD_BINARY(pow)
+    D2_STD_UNARY(tan)
     /*!
      * @brief `a[i] < b[i] ? *this[i] : other[i]`
      */
-    [[nodiscard]] constexpr Avec iflt(Avec other, Avec a, Avec b) const noexcept
+    [[nodiscard]] constexpr D2 iflt(D2 other, D2 a, D2 b) const noexcept
     {
-        return transform([a, b, other](T t, std::size_t i) noexcept
-                         { return a[i] < b[i] ? t : other.data[i]; });
+        return transform([a, b, other](std::size_t i, value_t t) noexcept
+                         { return a.data[i] < b.data[i] ? t : other.data[i]; });
     }
     /*!
      * @brief Get squared
      */
-    [[nodiscard]] constexpr Avec sq() const noexcept
+    [[nodiscard]] constexpr D2 sq() const noexcept
     {
-        Avec a{ *this };
+        D2 a{ *this };
         return a * a;
     }
     /*!
      * @brief Get cubed
      */
-    [[nodiscard]] constexpr Avec cb() const noexcept
+    [[nodiscard]] constexpr D2 cb() const noexcept
     {
-        Avec a{ *this };
+        D2 a{ *this };
         return a * a * a;
     }
 private:
@@ -131,8 +129,8 @@ private:
      * @brief Nullary transform
      */
     template<typename NullaryOp>
-        requires std::is_invocable_r_v<T, NullaryOp>
-    constexpr Avec transform(NullaryOp nullaryOp) const noexcept
+        requires std::is_invocable_r_v<value_t, NullaryOp>
+    constexpr D2 transform(NullaryOp nullaryOp) const noexcept
     {
         return transform([nullaryOp]([[maybe_unused]] auto i,
                                      [[maybe_unused]] auto t) noexcept
@@ -142,8 +140,8 @@ private:
      * @brief Unary transform
      */
     template<typename UnaryOp>
-        requires std::is_invocable_r_v<T, UnaryOp, T>
-    constexpr Avec transform(UnaryOp unaryOp) const noexcept
+        requires std::is_invocable_r_v<value_t, UnaryOp, value_t>
+    constexpr D2 transform(UnaryOp unaryOp) const noexcept
     {
         return transform([unaryOp]([[maybe_unused]] auto i, auto t) noexcept
                          { return unaryOp(t); });
@@ -152,8 +150,9 @@ private:
      * @brief Binary transform
      */
     template<typename BinaryOp>
-        requires std::is_nothrow_invocable_r_v<T, BinaryOp, T, T>
-    constexpr Avec transform(Avec other, BinaryOp binaryOp) const noexcept
+        requires std::
+            is_nothrow_invocable_r_v<value_t, BinaryOp, value_t, value_t>
+        constexpr D2 transform(D2 other, BinaryOp binaryOp) const noexcept
     {
         return transform([binaryOp, other](auto i, auto t) noexcept
                          { return binaryOp(t, other.data[i]); });
@@ -165,13 +164,15 @@ private:
      * @param enumeratedTransform
      */
     template<typename EnumeratedTransform>
-        requires std::
-            is_nothrow_invocable_r_v<T, EnumeratedTransform, std::size_t, T>
-        constexpr Avec
-        transform(EnumeratedTransform enumeratedTransform) const noexcept
+        requires std::is_nothrow_invocable_r_v<value_t,
+                                               EnumeratedTransform,
+                                               std::size_t,
+                                               value_t>
+    constexpr D2
+    transform(EnumeratedTransform enumeratedTransform) const noexcept
     {
-        Avec retval{};
-        for (std::size_t i = 0; i < n; ++i)
+        D2 retval{};
+        for (std::size_t i = 0; i < size_v; ++i)
         {
             retval.data[i] = enumeratedTransform(i, data[i]);
         }
@@ -180,10 +181,10 @@ private:
     /*!
      * @brief `std::array` instance
      */
-    alignas(sizeof(T) * n) std::array<T, n> data{};
+    alignas(sizeof(value_t) * size_v) std::array<value_t, size_v> data{};
 };
-} // namespace thezhe::avec
-namespace thezhe::avec::preconditions
+} // namespace thezhe::d2
+namespace thezhe::d2::preconditions
 {
 template<typename T>
 constexpr bool is_entirely_aligned_v{ sizeof(T) == alignof(T) };
@@ -198,19 +199,18 @@ constexpr bool is_nothrow_implicit_v{ std::is_nothrow_default_constructible_v<T>
                                       && std::is_nothrow_destructible_v<T>
                                       && !std::has_virtual_destructor_v<T> };
 /*!
- * @brief Validates an `Avec` specialization
+ * @brief Validates an `D2` specialization
  * @note `alignof(T) == alignof(typename T::data)`
- * @tparam T An `Avec` specialization
+ * @tparam T An `D2` specialization
  */
 template<typename T>
 constexpr bool is_valid_avec_v{ is_entirely_aligned_v<T>
                                 && is_128_bit_multiple_v<T>
                                 && is_nothrow_implicit_v<T> };
 template<typename T>
-constexpr bool is_valid_avec_128_v{ is_valid_avec_v<T> && sizeof(T) == 16 };
+constexpr bool is_valid_d2_v{ is_valid_avec_v<T> && sizeof(T) == 16 };
 /*!
  * @brief 128-bit simd must be valid
  */
-static_assert(is_valid_avec_128_v<Avec<double, 2>>);
-static_assert(is_valid_avec_128_v<Avec<float, 4>>);
-} // namespace thezhe::avec::preconditions
+static_assert(is_valid_d2_v<::thezhe::d2::D2>);
+} // namespace thezhe::d2::preconditions