+++ main-test.cpp
@@ -1,29 +1,32 @@
 #include <cstdlib>
-#include <ranges> // IWYU pragma: keep
-#include <thezhe/avec.hpp>
+#include <thezhe/d2.hpp>
 // TODO "CXXFLAGS": "-fopt-info-missed=missed.txt"
 // TODO compiler explorer see disassembly
-using Avec = thezhe::avec::Avec<>; // NOLINT
+using D2 = thezhe::d2::D2;
 void test_ctors() {}
 void test_assignment() {}
 void test_arithmetic() {}
 void test_math() {}
 int main()
 {
+    test_ctors();
+    test_assignment();
+    test_arithmetic();
+    test_math();
     // Template specialization and object
-    constexpr Avec avec{ 1.0, 1.0 };
+    constexpr D2 d2{ 1.0, 1.0 };
     // Constructors
-    constexpr Avec avec_copy_ctor{ avec };
-    static_assert(avec_copy_ctor == avec);
+    constexpr D2 avec_copy_ctor{ d2 };
+    static_assert(avec_copy_ctor == d2);
     // Assignment operators
-    Avec avec_copy{};
-    avec_copy = avec;
-    static_assert(avec_copy_ctor == avec);
+    D2 avec_copy{};
+    avec_copy = d2;
+    static_assert(avec_copy_ctor == d2);
     avec_copy += 1;
     // Assignment operators
     // Arithmetic operators
-    // static_assert(Avec(2.0, 4.0) == avec_copy + avec_copy);
-    // avec128d_copied = avec128d_copied * avec;
+    // static_assert(D2(2.0, 4.0) == avec_copy + avec_copy);
+    // avec128d_copied = avec128d_copied * d2;
     // auto test = -avec128d_copied;
     // Intrinsic math functions
     return EXIT_SUCCESS;